Legion_Omega_V0.33_ALPHA/requirements.txt
python-telegram-bot>=20.0
google-generativeai>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
Pillow>=10.0.0
mistralai>=1.0.0
httpx>=0.25.0
